moved App to Home folder
