+++ src/Home/App.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import logo from "./logo.svg";
-import { Counter } from "./features/counter/Counter";
-import { Form } from "./features/form";
+import { Counter } from "../features/counter/Counter";
+import { Form } from "../features/form";
 import { useAuth0 } from "@auth0/auth0-react";
 import "./App.css";
 
@@ -18,9 +18,20 @@ function App() {
           <button onClick={() => logout({ returnTo: window.location.origin })}>
             Log Out
           </button>
-          <p>{user?.email}</p>
-          <h1>Register and add your wishlist</h1>
-          <Form />
+          {/* Conditionally render if exchanges exist */}
+          <div>
+            <p>Your gift exchanges:</p>
+            {/* Map exchanges out as ul of a tags */}
+            <a href={`exchange`}>Exchange xyz</a>
+          </div>
+          <div>
+            <p>Join an exchange:</p>
+            <a href={`find`}>Find</a>
+          </div>
+          <div>
+            <p>Create a gift exchange:</p>
+            <a href={`find`}>Create</a>
+          </div>
         </>
       )}
     </div>
