+++ src/index.tsx
@@ -5,7 +5,7 @@ import { Auth0Provider } from "@auth0/auth0-react";
 import { createBrowserRouter, RouterProvider, Route } from "react-router-dom";
 
 import { store } from "./app/store";
-import App from "./App";
+import App from "./Home/App";
 import reportWebVitals from "./reportWebVitals";
 import "./index.css";
